complete other apps function.

+++ article/views.py
@@ -4,6 +4,8 @@ import json
 from django.core import serializers
 from django.views.decorators.csrf import csrf_exempt
 from article.models import Article
+from user.models import User
+from role.models import Role
 
 # Create your views here.
 
@@ -39,7 +41,10 @@ def delete_by_id(request):
     print(id)
     print(user_id)
     result = Article.objects.get(id=id)
-    if result.user == int(user_id):
+    user_role = User.objects.get(id=user_id).role
+    ok_del = Role.objects.get(id=user_role).del_article
+
+    if result.user == int(user_id) and ok_del:
         response_data = {'status': '200', 'message': '成功。', 'result': '操作成功。'}
         return HttpResponse(json.dumps(response_data), content_type="application/json")
         result.delete()
@@ -49,21 +54,53 @@ def delete_by_id(request):
 
 
 def select_by_id(request):
-    id = request.GET.get('id')
-    result = Article.objects.get(id=id)
-    return HttpResponse(result, content_type="application/json")
+    try:
+        id = request.GET.get('id')
+        result = Article.objects.get(id=id)
+        return HttpResponse(result, content_type="application/json")
 
-
+    except:
+        response_data = {'status': '404', 'message': '失败：不存在。', 'result': 'Data Not Found'}
+        return HttpResponse(json.dumps(response_data), content_type="application/json")
 
 def select_all(request):
-    result = Article.objects.all()
-    result = serializers.serialize("json", result)
+    try:
+        result = Article.objects.all()
+        result = serializers.serialize("json", result)
+        return HttpResponse(result, content_type="application/json")
+    except:
+        response_data = {'status': '404', 'message': '失败：不存在。', 'result': 'Data Not Found'}
+        return HttpResponse(json.dumps(response_data), content_type="application/json")
 
-    return HttpResponse(result, content_type="application/json")
 
+@csrf_exempt
+def update_by_id(request):
+    print(request.method)
+    if request.method != 'POST':
+        response_data = {'status': '405', 'message': '失败：请求方式错误，请使用POST方式', 'result': 'Method Not Allowed'}
+        return HttpResponse(json.dumps(response_data), content_type="application/json")
+    param = json.loads(request.body.decode('utf-8'))
 
-def update_by_id(request, id):
-    if request.POST:
-        res = Article.objects.get(id=str(id))
-        print(res)
-        return HttpResponse("Hello Article")
+    try:
+        id = param['id']
+        title = param['title']
+        content = param['content']
+        user = param['user']
+
+        result = Article.objects.get(id=id)
+        user_role = User.objects.get(id=user).role
+        ok_edit = Role.objects.get(id=user_role).edit_article
+
+        if result.user == int(user) and ok_edit:
+            result.title = title
+            result.content = content
+            result.save()
+            response_data = {'status': '200', 'message': '成功', 'result': title}
+            return HttpResponse(json.dumps(response_data), content_type="application/json")
+        else:
+            response_data = {'status': '407', 'message': '失败：用户权限不足。', 'result': 'Opened Not Allowed'}
+            return HttpResponse(json.dumps(response_data), content_type="application/json")
+
+    except ValueError:
+        response_data = {'status': '406', 'message': '失败：参数设置错误，请检查参数类型。', 'result': 'Method Not Allowed'}
+        return HttpResponse(json.dumps(response_data), content_type="application/json")
